Add new keyring, fix node executor tests
