+++ core/keyring/src/lib.rs
@@ -16,158 +16,13 @@
 
 //! Support code for the runtime. A set of test accounts.
 
-use std::collections::HashMap;
-use std::ops::Deref;
-use lazy_static::lazy_static;
-use hex_literal::{hex, hex_impl};
-use substrate_primitives::ed25519::{Pair, Public, Signature};
-pub use substrate_primitives::ed25519;
+/// Test account crypto for sr25519.
+pub mod sr25519;
 
-/// Set of test accounts.
-#[derive(Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Keyring {
-	Alice,
-	Bob,
-	Charlie,
-	Dave,
-	Eve,
-	Ferdie,
-	One,
-	Two,
-}
+/// Test account crypto for ed25519.
+pub mod ed25519;
 
-impl Keyring {
-	pub fn from_public(who: Public) -> Option<Keyring> {
-		[
-			Keyring::Alice,
-			Keyring::Bob,
-			Keyring::Charlie,
-			Keyring::Dave,
-			Keyring::Eve,
-			Keyring::Ferdie,
-			Keyring::One,
-			Keyring::Two,
-		].iter()
-			.map(|i| *i)
-			.find(|&k| Public::from(k) == who)
-	}
-
-	pub fn from_raw_public(who: [u8; 32]) -> Option<Keyring> {
-		Self::from_public(Public::from_raw(who))
-	}
-
-	pub fn to_raw_public(self) -> [u8; 32] {
-		*Public::from(self).as_array_ref()
-	}
-
-	pub fn to_raw_public_vec(self) -> Vec<u8> {
-		Public::from(self).to_raw_vec()
-	}
-
-	pub fn sign(self, msg: &[u8]) -> Signature {
-		Pair::from(self).sign(msg)
-	}
-
-	pub fn pair(self) -> Pair {
-		match self {
-			Keyring::Alice => Pair::from_seed(b"Alice                           "),
-			Keyring::Bob => Pair::from_seed(b"Bob                             "),
-			Keyring::Charlie => Pair::from_seed(b"Charlie                         "),
-			Keyring::Dave => Pair::from_seed(b"Dave                            "),
-			Keyring::Eve => Pair::from_seed(b"Eve                             "),
-			Keyring::Ferdie => Pair::from_seed(b"Ferdie                          "),
-			Keyring::One => Pair::from_seed(b"12345678901234567890123456789012"),
-			Keyring::Two => Pair::from_seed(&hex!("9d61b19deffd5a60ba844af492ec2cc44449c5697b326919703bac031cae7f60")),
-		}
-	}
-}
-
-impl From<Keyring> for &'static str {
-	fn from(k: Keyring) -> Self {
-		match k {
-			Keyring::Alice => "Alice",
-			Keyring::Bob => "Bob",
-			Keyring::Charlie => "Charlie",
-			Keyring::Dave => "Dave",
-			Keyring::Eve => "Eve",
-			Keyring::Ferdie => "Ferdie",
-			Keyring::One => "one",
-			Keyring::Two => "two",
-		}
-	}
-}
-
-lazy_static! {
-	static ref PRIVATE_KEYS: HashMap<Keyring, Pair> = {
-		[
-			Keyring::Alice,
-			Keyring::Bob,
-			Keyring::Charlie,
-			Keyring::Dave,
-			Keyring::Eve,
-			Keyring::Ferdie,
-			Keyring::One,
-			Keyring::Two,
-		].iter().map(|&i| (i, i.pair())).collect()
-	};
-
-	static ref PUBLIC_KEYS: HashMap<Keyring, Public> = {
-		PRIVATE_KEYS.iter().map(|(&name, pair)| (name, pair.public())).collect()
-	};
-}
-
-impl From<Keyring> for Public {
-	fn from(k: Keyring) -> Self {
-		(*PUBLIC_KEYS).get(&k).unwrap().clone()
-	}
-}
-
-impl From<Keyring> for Pair {
-	fn from(k: Keyring) -> Self {
-		k.pair()
-	}
-}
-
-impl From<Keyring> for [u8; 32] {
-	fn from(k: Keyring) -> Self {
-		*(*PUBLIC_KEYS).get(&k).unwrap().as_array_ref()
-	}
-}
-
-impl From<Keyring> for &'static [u8; 32] {
-	fn from(k: Keyring) -> Self {
-		(*PUBLIC_KEYS).get(&k).unwrap().as_array_ref()
-	}
-}
-
-impl AsRef<[u8; 32]> for Keyring {
-	fn as_ref(&self) -> &[u8; 32] {
-		(*PUBLIC_KEYS).get(self).unwrap().as_array_ref()
-	}
-}
-
-impl AsRef<Public> for Keyring {
-	fn as_ref(&self) -> &Public {
-		(*PUBLIC_KEYS).get(self).unwrap()
-	}
-}
-
-impl Deref for Keyring {
-	type Target = [u8; 32];
-	fn deref(&self) -> &[u8; 32] {
-		(*PUBLIC_KEYS).get(self).unwrap().as_array_ref()
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use ed25519::Verifiable;
-
-	#[test]
-	fn should_work() {
-		assert!(Keyring::Alice.sign(b"I am Alice!").verify(b"I am Alice!", Keyring::Alice));
-		assert!(!Keyring::Alice.sign(b"I am Alice!").verify(b"I am Bob!", Keyring::Alice));
-		assert!(!Keyring::Alice.sign(b"I am Alice!").verify(b"I am Alice!", Keyring::Bob));
-	}
-}
+/// The Ed25519 keyring.
+/// 
+/// This is deprecated: use `ed25519::Keyring` instead.
+pub use ed25519::Keyring;
